refactor(new recipe): move files
