+++ src/app/components/NewRecipe/components/SearchProductModal/ExistingProduct.tsx
