+++ src/app/components/NewRecipe/components/SearchProductModal/NewProduct.tsx
